Refactor `ConditionMigration`
